agents/chat/requirements.txt
fastapi
pydantic
uvicorn
openai
sentence-transformers
scikit-learn
google-cloud-secret-manager
python-dotenv
tiktoken
langchain
langchain-openai
faiss-cpu 